feat: Create EXCSLA.Shared.Core project and migrate base domain classes
