+++ src/Core/EXCSLA.Shared.Core/IgnoreMemberAttribute.cs
